Fix add for non admin users in closed
I have checked that non-admin users of closed groups can not create a webinar using canEdit(). With this fix in line 97 looks like access level are correctly checked.
I look through the mod/groups/lib/discussions.php to look for a solution and found the use of canWriteToContainer as of http://reference.elgg.org/classElggEntity.html#ae127fead8cba88c93136429f1aa2cb86
so it checks if it is possible to edit a webinar in the group, instead of checking with canEdit if you can edit the group.